import 1.0.8
2003-09-22  James Antill  <james@and.org>

	* NEWS: Release-1.0.8

2003-09-20  James Antill  <james@and.org>

	* src/vstr_sc.c (vstr__sc_fmt_num_ipv6): Fix couple of bugs for length
	ipv6 output with IPV4 at the end.

2003-09-18  James Antill  <james@and.org>

	* include/vstr-switch.h (VSTR__COMPILE_ATTR_WARN_UNUSED_RET): Add
	warn_unused_return gcc compiler attribute.

	* include/vstr-extern.h (vstr_dup_rep_chr): Remove nonnull attribute.

	* src/vstr_cntl.c (vstr_cntl_opt): Fix fallthrough error when compiling
	in debug mode.

	* src/vstr_ref.c (vstr_ref_make_memdup): Add test for malloc failure.
	(vstr_ref_make_memdup): Change pointer to const, this is ABI compatible.
	(vstr_ref_make_ptr): Change pointer to const, this is ABI compatible.

2003-09-12  James Antill  <james@and.org>

	* include/vstr-inline.h (vstr_ref_make_strdup): Added function.

2003-09-02  James Antill  <james@and.org>

	* src/vstr_parse.c (vstr_parse_ipv4): Minor cidr wasn't written if
	ip was entire size of section.

2003-08-23  James Antill  <james@and.org>

	* Documentation/security_problems.txt: Added more
	classifications

git-archimport-id: james@and.org--2004-code/vstr--main--1--patch-38

+++ examples/ex_perf.h
@@ -71,7 +71,7 @@
                tv_end.tv_usec - tv_beg.tv_usec, \
                ((long)mal_end.uordblks - (long)mal_beg.uordblks) + \
                ((long)mal_end.hblkhd - (long)mal_beg.hblkhd), \
-               (mal_end.fordblks - mal_beg.fordblks)); \
+               ((long)mal_end.fordblks - (long)mal_beg.fordblks)); \
 	} while (0)
